PR comments fixes: * serverSentEventBody -> serverSentEventsBody * Using actorSystem.dispatcher instead of global EC * Moving code related to SSE into separate files
