+++ server/akka-http-server/src/main/scala/sttp/tapir/server/akkahttp/AkkaHttpServerInterpreter.scala
@@ -1,12 +1,8 @@
 package sttp.tapir.server.akkahttp
 
-import akka.NotUsed
 import akka.http.scaladsl.server.{Directive, Route}
-import akka.stream.scaladsl.{Flow, Framing, Source}
-import akka.util.ByteString
 import sttp.capabilities.WebSockets
 import sttp.capabilities.akka.AkkaStreams
-import sttp.model.sse.ServerSentEvent
 import sttp.tapir.Endpoint
 import sttp.tapir.server.ServerEndpoint
 
@@ -48,28 +44,6 @@ trait AkkaHttpServerInterpreter {
   ): Route = {
     new EndpointToAkkaServer(serverOptions).toRoute(serverEndpoints)
   }
-
-  def serialiseSSEToBytes: Source[ServerSentEvent, Any] => AkkaStreams.BinaryStream = sseStream =>
-    sseStream.map(sse => {
-      ByteString(s"${composeSSE(sse)}\n\n")
-    })
-
-  private def composeSSE(sse: ServerSentEvent) = {
-    val data = sse.data.map(_.split("\n")).map(_.map(line => Some(s"data: $line"))).getOrElse(Array.empty)
-    val event = sse.eventType.map(event => s"event: $event")
-    val id = sse.id.map(id => s"id: $id")
-    val retry = sse.retry.map(retryCount => s"retry: $retryCount")
-    (data :+ event :+ id :+ retry).flatten.mkString("\n")
-  }
-
-  def parseBytesToSSE: AkkaStreams.BinaryStream => Source[ServerSentEvent, Any] = stream => stream.via(parse)
-
-  private val parse: Flow[ByteString, ServerSentEvent, NotUsed] =
-    Framing
-      .delimiter(ByteString("\n\n"), maximumFrameLength = Int.MaxValue, allowTruncation = true)
-      .map(_.utf8String)
-      .map(_.split("\n").toList)
-      .map(ServerSentEvent.parse)
 }
 
 object AkkaHttpServerInterpreter extends AkkaHttpServerInterpreter
