+++ server/akka-http-server/src/main/scala/sttp/tapir/server/akkahttp/akkahttp.scala
@@ -9,7 +9,7 @@ import sttp.tapir.{CodecFormat, Schema, StreamBodyIO, streamBody}
 import java.nio.charset.Charset
 
 package object akkahttp extends TapirAkkaHttpServer {
-  val serverSentEventBody: StreamBodyIO[Source[ByteString, Any], Source[ServerSentEvent, Any], AkkaStreams] =
+  val serverSentEventsBody: StreamBodyIO[Source[ByteString, Any], Source[ServerSentEvent, Any], AkkaStreams] =
     streamBody(AkkaStreams)(Schema.binary, CodecFormat.TextEventStream(), Some(Charset.forName("UTF-8")))
-      .map(AkkaHttpServerInterpreter.parseBytesToSSE)(AkkaHttpServerInterpreter.serialiseSSEToBytes)
+      .map(AkkaServerSentEvents.parseBytesToSSE)(AkkaServerSentEvents.serialiseSSEToBytes)
 }
