+++ server/http4s-server/src/main/scala/sttp/tapir/server/http4s/Http4sServerInterpreter.scala
@@ -5,9 +5,7 @@ import cats.effect.{Concurrent, ContextShift, Sync, Timer}
 import cats.~>
 import org.http4s.{Http, HttpRoutes}
 import sttp.capabilities.WebSockets
-import fs2.{Stream, text}
 import sttp.capabilities.fs2.Fs2Streams
-import sttp.model.sse.ServerSentEvent
 import sttp.tapir.Endpoint
 import sttp.tapir.server.ServerEndpoint
 
@@ -86,32 +84,6 @@ trait Http4sServerInterpreter {
       timer: Timer[F]
   ): HttpRoutes[F] = {
     new EndpointToHttp4sServer(serverOptions).toRoutes(serverEndpoints)
-  }
-
-  def serialiseSSEToBytes[F[_]](streams: Fs2Streams[F]): Stream[F, ServerSentEvent] => streams.BinaryStream = sseStream => {
-    sseStream
-      .map(sse => {
-        s"${composeSSE(sse)}\n\n"
-      })
-      .through(text.utf8Encode)
-  }
-
-  private def composeSSE(sse: ServerSentEvent) = {
-    val data = sse.data.map(_.split("\n")).map(_.map(line => Some(s"data: $line"))).getOrElse(Array.empty)
-    val event = sse.eventType.map(event => s"event: $event")
-    val id = sse.id.map(id => s"id: $id")
-    val retry = sse.retry.map(retryCount => s"retry: $retryCount")
-    (data :+ event :+ id :+ retry).flatten.mkString("\n")
-  }
-
-  def parseBytesToSSE[F[_]](streams: Fs2Streams[F]): streams.BinaryStream => Stream[F, ServerSentEvent] = stream => {
-    stream
-      .through(text.utf8Decode[F])
-      .through(text.lines[F])
-      .split(_.isEmpty)
-      .filter(_.nonEmpty)
-      .map(_.toList)
-      .map(ServerSentEvent.parse)
   }
 }
