+++ server/http4s-server/src/main/scala/sttp/tapir/server/http4s/http4s.scala
@@ -7,9 +7,9 @@ import sttp.tapir.{CodecFormat, Schema, StreamBodyIO, streamBody}
 import java.nio.charset.Charset
 
 package object http4s extends TapirHttp4sServer {
-  def serverSentEventBody[F[_]]: StreamBodyIO[fs2.Stream[F, Byte], fs2.Stream[F, ServerSentEvent], Fs2Streams[F]] = {
+  def serverSentEventsBody[F[_]]: StreamBodyIO[fs2.Stream[F, Byte], fs2.Stream[F, ServerSentEvent], Fs2Streams[F]] = {
     val fs2Streams = Fs2Streams[F]
     streamBody(fs2Streams)(Schema.binary, CodecFormat.TextEventStream(), Some(Charset.forName("UTF-8")))
-      .map(Http4sServerInterpreter.parseBytesToSSE(fs2Streams))(Http4sServerInterpreter.serialiseSSEToBytes(fs2Streams))
+      .map(Http4sServerSentEvents.parseBytesToSSE(fs2Streams))(Http4sServerSentEvents.serialiseSSEToBytes(fs2Streams))
   }
 }
